gcp/requirements.txt
functions-framework==3.5.0
tensorflow==2.15.0
Pillow==10.2.0
numpy==1.26.4
google-cloud-storage==2.14.0